Fix removal in identity views and widen hill-climber counters
values().remove(o): for weak/soft value caches now uses identity
equivalence (matching IdentityHashMap's semantics) rather than Guava's
accidental equals-based behavior (inherited from AbstractCollection).

keySet()/entrySet().removeAll: the size optimization picks between
iterating the cache vs. iterating the user collection, but for weak
keys the cache compares by identity. Iterating the user collection
calls cache.remove(item) with a possibly equal-but-distinct instance,
which misses by identity. Always iterate the cache when collectKeys()
is true, matching IdentityHashMap.KeySet.removeAll's defensive comment.

hitsInSample and missesInSample widen from int to long to avoid
counters overflowing before the hill climber reset triggered, leaving
requestCount stuck negative and permanently disabling adjustment.

Expanded the internal consistency checks in the test assertions for a
more thorough verification.

+++ caffeine/src/javaPoet/java/com/github/benmanes/caffeine/cache/local/AddMaximum.java
@@ -70,8 +70,8 @@ public final class AddMaximum implements Rule<LocalCacheContext> {
   private static void addHillClimber(LocalCacheContext context) {
     addField(context, double.class, "stepSize");
     addField(context, long.class, "adjustment");
-    addField(context, int.class, "hitsInSample");
-    addField(context, int.class, "missesInSample");
+    addField(context, long.class, "hitsInSample");
+    addField(context, long.class, "missesInSample");
     addField(context, double.class, "previousSampleHitRate");
   }
